Wordle Game
This is a web-based implementation of the popular guessing game, Wordle, created using the React framework. The game challenges players to guess a five-letter word within a certain number of guesses, with each correct guess revealed in its correct position and each incorrect guess represented with a symbol. The goal of the game is to guess the word in as few attempts as possible.